make parent child routing
